Сделанная домашняя работа по 22-ому спринту

+++ app/src/main/java/com/practicum/playlistmaker/library/domain/api/repo/LibraryDbRepository.kt
@@ -8,5 +8,5 @@ interface LibraryDbRepository {
     suspend fun getTracksFromFavoriteRepo(): Flow<List<Track>>
     suspend fun insertTrackToFavoriteRepo(track: Track)
     suspend fun deleteTrackFromFavoriteRepo(track: Track)
-    suspend fun isTrackInFavorites(trackId: Int): Boolean
+    suspend fun isTrackInFavorites(trackId: Long): Boolean
 }
