+++ app/src/main/java/com/practicum/playlistmaker/search/domain/models/TrackInterface.kt
@@ -5,7 +5,7 @@ interface TrackInterface {
     val artistName: String
     val trackTimeMillis: Long
     val artworkUrl100: String
-    val trackId: Int
+    val trackId: Long
     val collectionName: String?
     val releaseDate: String
     val primaryGenreName: String
